refactor: Dealer는 클래스 제거에 따른 코드 수정
- Player와 Dealer의 행동은 다르지 않다.
- 따라서 Dealer 클래스를 삭제한다.
- 달라야하는 기능이 발견된다면 Dealer를 다시 만들고, 조합으로 해결한다.
